+++ lib/chef_gen/flavor.rb
@@ -1,0 +1,5 @@
+# chef generators
+module ChefGen
+  # a namespace for template flavors
+  module Flavor; end
+end