+++ spec/support/fixtures/lib/chef-dk/template/plugin/baz.rb
@@ -1,8 +1,0 @@
-module ChefDK
-  module Template
-    class Plugin
-      class Baz
-      end
-    end
-  end
-end