+++ spec/support/fixtures/lib/chef_gen/flavor/bar.rb
@@ -1,0 +1,21 @@
+module ChefGen
+  module Flavor
+    class Bar
+      class << self
+        def description
+          'bar cookbook template'
+        end
+
+        def code_generator_path(classfile)
+          File.expand_path(
+            File.join(
+              classfile,
+              '..', '..', '..', '..',
+              'code_generator_2'
+            )
+          )
+        end
+      end
+    end
+  end
+end