+++ spec/support/fixtures/lib/chef_gen/flavor/foo.rb
@@ -1,0 +1,9 @@
+module ChefGen
+  module Flavor
+    class Foo
+      def self.description
+        'foo cookbook template'
+      end
+    end
+  end
+end